requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
reportlab==4.0.7
pydantic==1.10.14
